Part 4 Blog List Expansions step 9

+++ Part4/Blog_List/controllers/blogs.js
@@ -14,9 +14,7 @@ BlogsRouter.get('/', async (request, response) => {
 BlogsRouter.post('/', async (request, response) => {
     let blog = new Blog(request.body)
     const decodedToken = jwt.verify(request.token, config.JWT_SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
+    if (!decodedToken.id) return response.status(401).json({ error: 'token invalid' })
     let authorUser = await User.findOne({ _id: decodedToken.id })
     blog.User = authorUser.id
     const result = await blog.save()
@@ -26,8 +24,17 @@ BlogsRouter.post('/', async (request, response) => {
 })
 
 BlogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, config.JWT_SECRET)
+    if (!decodedToken.id) return response.status(401).json({ error: 'token invalid' })
+    const user = await User.findOne({ _id: decodedToken.id })
+    if (!user) return response.status(401).json({ error: 'no valid user' })
+    const blog = await Blog.findOne({ _id: request.params.id })
+    if (!blog) return response.status(404).json({ error: 'no valid blog' })
+    if (blog.User.toString() !== decodedToken.id) return response.status(401).json({ error: 'You dont have permission to delete this blog' })
     await Blog.deleteOne({ _id: request.params.id })
-    response.status(201).end()
+    user.Blog = user.Blog.filter((e) => (e.toString !== blog.id))
+    await user.save()
+    response.status(204).end()
 })
 
 BlogsRouter.put('/:id', async (request, response) => {
